fix(types): taro-service 使用项目生成的类型
