+++ packages/taro-service/package.json
@@ -3,7 +3,7 @@
   "version": "3.5.7",
   "description": "Taro Service",
   "main": "index.js",
-  "types": "types/index.d.ts",
+  "types": "dist/index.d.ts",
   "scripts": {
     "build": "run-s clean prod",
     "dev": "tsc -w",
