+++ packages/taro-service/src/index.ts
@@ -4,3 +4,5 @@ import { TaroPlatformBase } from './platform-plugin-base'
 export { Kernel }
 export { TaroPlatformBase }
 export default { Kernel, TaroPlatformBase }
+
+export type { IPluginContext } from './utils/types'
